+++ gradle.properties
@@ -1,5 +1,5 @@
 groupId=ru.tinkoff.kora
-koraVersion=1.0.9
+koraVersion=1.1.0-SNAPSHOT
 
 
 ##### GRADLE #####
