+++ kora-java-graalvm-crud-jdbc/build.gradle
@@ -1,0 +1,182 @@
+buildscript {
+    dependencies {
+        classpath("ru.tinkoff.kora:openapi-generator:$koraVersion")
+    }
+}
+
+plugins {
+    id "java"
+    id "jacoco"
+    id "application"
+
+    id "org.openapi.generator" version "7.1.0"
+    id "com.github.johnrengelman.shadow" version "8.1.1"
+    id "org.flywaydb.flyway" version "8.4.2"
+    id "org.graalvm.buildtools.native" version "0.10.1"
+}
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+}
+
+mainClassName = "ru.tinkoff.kora.example.graalvm.crud.jdbc.Application"
+
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17
+
+configurations {
+    koraBom
+    implementation.extendsFrom(koraBom)
+    annotationProcessor.extendsFrom(koraBom)
+}
+
+dependencies {
+    koraBom platform("ru.tinkoff.kora:kora-parent:$koraVersion")
+    annotationProcessor "org.mapstruct:mapstruct-processor:1.5.5.Final"
+    annotationProcessor "ru.tinkoff.kora:annotation-processors"
+    annotationProcessor "org.slf4j:slf4j-simple:2.0.11"
+    annotationProcessor "io.goodforgod:graalvm-hint-processor:1.1.1"
+    compileOnly "io.goodforgod:graalvm-hint-annotations:1.1.1"
+
+    implementation "ru.tinkoff.kora:http-server-undertow"
+    implementation "ru.tinkoff.kora:database-jdbc"
+    implementation "ru.tinkoff.kora:micrometer-module"
+    implementation "ru.tinkoff.kora:json-module"
+    implementation "ru.tinkoff.kora:validation-module"
+    implementation "ru.tinkoff.kora:validation-common"
+    implementation "ru.tinkoff.kora:cache-caffeine"
+    implementation "ru.tinkoff.kora:resilient-kora"
+    implementation "ru.tinkoff.kora:config-hocon"
+    implementation "ru.tinkoff.kora:openapi-management"
+    implementation "ru.tinkoff.kora:logging-logback"
+
+    runtimeOnly "org.postgresql:postgresql:42.6.0"
+    implementation "org.mapstruct:mapstruct:1.5.5.Final"
+
+    testImplementation "org.json:json:20231013"
+    testImplementation "org.skyscreamer:jsonassert:1.5.1"
+
+    testImplementation "org.mockito:mockito-core:5.6.0"
+    testImplementation "ru.tinkoff.kora:test-junit5"
+    testImplementation "io.goodforgod:testcontainers-extensions-postgres:0.9.5"
+    testImplementation "org.testcontainers:junit-jupiter:1.17.6"
+}
+
+openApiGenerate {
+    generatorName = "kora"
+    group = "openapi tools"
+    inputSpec = "$projectDir/src/main/resources/openapi/http-server.yaml"
+    outputDir = "$buildDir/generated/openapi"
+    apiPackage = "ru.tinkoff.kora.example.graalvm.crud.openapi.server.api"
+    modelPackage = "ru.tinkoff.kora.example.graalvm.crud.openapi.server.model"
+    invokerPackage = "ru.tinkoff.kora.example.graalvm.crud.openapi.server.invoker"
+    configOptions = [
+            mode                  : "java-server", // так же есть java-server вариация HTTP Server"а
+            enableServerValidation: "true"
+    ]
+}
+
+graalvmNative {
+    binaries {
+        main {
+            imageName = "$project.name"
+            mainClass = "$mainClassName"
+            javaLauncher = javaToolchains.launcherFor {
+                languageVersion = JavaLanguageVersion.of(21)
+                vendor = JvmVendorSpec.matching("GraalVM Community")
+            }
+        }
+    }
+    metadataRepository {
+        enabled = true
+    }
+}
+
+compileJava.dependsOn tasks.openApiGenerate
+//processResources.dependsOn tasks.collectReachabilityMetadata
+test.dependsOn tasks.shadowJar
+
+//noinspection GroovyAssignabilityCheck
+run {
+    environment([
+            "POSTGRES_JDBC_URL": "jdbc:postgresql://$postgresHost:$postgresPort/$postgresDatabase",
+            "POSTGRES_USER"    : "$postgresUser",
+            "POSTGRES_PASS"    : "$postgresPassword",
+    ])
+}
+
+test {
+    jvmArgs += [
+            "-XX:+TieredCompilation",
+            "-XX:TieredStopAtLevel=1",
+    ]
+
+    environment([
+            "": ""
+    ])
+
+    useJUnitPlatform()
+    testLogging {
+        showStandardStreams(true)
+        events("passed", "skipped", "failed")
+        exceptionFormat("full")
+    }
+
+    jacoco {
+        excludes += ["**/Application*"]
+    }
+
+    reports {
+        html.required = false
+        junitXml.required = false
+    }
+}
+
+sourceSets {
+    main {
+        java.srcDirs += "$buildDir/generated/openapi"
+        resources.srcDirs += "$buildDir/native-reachability-metadata"
+    }
+}
+
+flyway {
+    url = "jdbc:postgresql://$postgresHost:$postgresPort/$postgresDatabase"
+    user = "$postgresUser"
+    password = "$postgresPassword"
+    locations = ["classpath:db/migration"]
+}
+
+jar.enabled = false
+shadowJar {
+    mergeServiceFiles()
+    manifest {
+        attributes "Main-Class": mainClassName
+        attributes "Implementation-Version": koraVersion
+    }
+}
+
+artifacts {
+    archives shadowJar
+}
+
+compileJava {
+    options.encoding("UTF-8")
+    options.incremental(true)
+    options.fork = true
+}
+
+check.dependsOn jacocoTestReport
+jacocoTestReport {
+    reports {
+        xml.required = true
+        html.outputLocation = layout.buildDirectory.dir("jacocoHtml")
+    }
+}
+
+javadoc {
+    options.encoding = "UTF-8"
+    if (JavaVersion.current().isJava9Compatible()) {
+        options.addBooleanOption("html5", true)
+    }
+}
