+++ kora-java-graalvm-crud-jdbc/gradle.properties
