+++ kora-java-graalvm-crud-jdbc/src/main/java/ru/tinkoff/kora/example/graalvm/crud/jdbc/model/dao/PetCategory.java
@@ -1,0 +1,8 @@
+package ru.tinkoff.kora.example.graalvm.crud.jdbc.model.dao;
+
+import ru.tinkoff.kora.database.common.annotation.Id;
+import ru.tinkoff.kora.database.common.annotation.Table;
+
+@Table("categories")
+public record PetCategory(@Id long id,
+                          String name) {}
