+++ kora-java-graalvm-crud-jdbc/src/main/java/ru/tinkoff/kora/example/graalvm/crud/jdbc/service/PetCache.java
@@ -1,0 +1,10 @@
+package ru.tinkoff.kora.example.graalvm.crud.jdbc.service;
+
+import ru.tinkoff.kora.cache.annotation.Cache;
+import ru.tinkoff.kora.cache.caffeine.CaffeineCache;
+import ru.tinkoff.kora.example.graalvm.crud.jdbc.model.dao.PetWithCategory;
+
+@Cache("pet-cache")
+public interface PetCache extends CaffeineCache<Long, PetWithCategory> {
+
+}
