+++ kora-java-graalvm-crud-jdbc/src/main/resources/application.conf
@@ -1,0 +1,77 @@
+httpServer {
+  publicApiHttpPort = 8080
+  privateApiHttpPort = 8085
+}
+
+
+db {
+  jdbcUrl = ${POSTGRES_JDBC_URL}
+  r2dbcUrl = ${POSTGRES_R2DBC_URL}
+  connectionUri = ${POSTGRES_VERTX_URL}
+  username = ${POSTGRES_USER}
+  password = ${POSTGRES_PASS}
+  maxPoolSize = 10
+  poolName = "kora"
+  initializationFailTimeout = "10s"
+}
+
+
+cassandra {
+  auth {
+    login = ${CASSANDRA_USER}
+    password = ${CASSANDRA_PASS}
+  }
+  basic {
+    contactPoints = ${CASSANDRA_CONTACT_POINTS}
+    dc = ${CASSANDRA_DC}
+    sessionKeyspace = ${CASSANDRA_KEYSPACE}
+    request {
+      timeout = 5s
+    }
+  }
+}
+
+pet-cache {
+  maximumSize = 1000
+  expireAfterWrite = ${?CACHE_EXPIRE_WRITE}
+}
+
+
+openapi {
+  management {
+    enabled = true
+    file = "openapi/http-server.yaml"
+    swaggerui {
+      enabled = true
+    }
+    rapidoc {
+      enabled = true
+    }
+  }
+}
+
+
+resilient {
+  circuitbreaker.pet {
+    slidingWindowSize = 50
+    minimumRequiredCalls = 25
+    failureRateThreshold = 50
+    permittedCallsInHalfOpenState = 10
+    waitDurationInOpenState = 15s
+  }
+  timeout.pet {
+    duration = 5000ms
+  }
+  retry.pet {
+    delay = 100ms
+    attempts = 2
+  }
+}
+
+
+logging.level {
+  "root": "WARN"
+  "ru.tinkoff.kora": "INFO"
+  "ru.tinkoff.kora.example": "INFO"
+  "ru.tinkoff.kora.application.graph.internal.loom.VirtualThreadExecutorHolder": "DEBUG"
+}
