+++ kora-java-graalvm-crud-jdbc/src/main/resources/db/migration/V1__setup-tables.sql
