--- kora-java-graalvm-crud/src/main/java/ru/tinkoff/kora/example/graalvm/crud/model/dao/PetCategory.java
@@ -1,8 +1,0 @@
-package ru.tinkoff.kora.example.graalvm.crud.model.dao;
-
-import ru.tinkoff.kora.database.common.annotation.Id;
-import ru.tinkoff.kora.database.common.annotation.Table;
-
-@Table("categories")
-public record PetCategory(@Id long id,
-                          String name) {}
