--- kora-java-graalvm-crud/src/main/java/ru/tinkoff/kora/example/graalvm/crud/service/PetCache.java
@@ -1,10 +1,0 @@
-package ru.tinkoff.kora.example.graalvm.crud.service;
-
-import ru.tinkoff.kora.cache.annotation.Cache;
-import ru.tinkoff.kora.cache.caffeine.CaffeineCache;
-import ru.tinkoff.kora.example.graalvm.crud.model.dao.PetWithCategory;
-
-@Cache("pet-cache")
-public interface PetCache extends CaffeineCache<Long, PetWithCategory> {
-
-}
