--- kora-java-graalvm-crud/src/main/resources/application.conf
@@ -1,60 +1,0 @@
-httpServer {
-  publicApiHttpPort = 8080
-  privateApiHttpPort = 8085
-}
-
-
-db {
-  jdbcUrl = ${POSTGRES_JDBC_URL}
-  username = ${POSTGRES_USER}
-  password = ${POSTGRES_PASS}
-  maxPoolSize = 10
-  poolName = "kora"
-  initializationFailTimeout = "10s"
-}
-
-
-pet-cache {
-  maximumSize = 1000
-  expireAfterWrite = ${?CACHE_EXPIRE_WRITE}
-}
-
-
-openapi {
-  management {
-    enabled = true
-    file = "openapi/http-server.yaml"
-    swaggerui {
-      enabled = true
-    }
-    rapidoc {
-      enabled = true
-    }
-  }
-}
-
-
-resilient {
-  circuitbreaker.pet {
-    slidingWindowSize = 50
-    minimumRequiredCalls = 25
-    failureRateThreshold = 50
-    permittedCallsInHalfOpenState = 10
-    waitDurationInOpenState = 15s
-  }
-  timeout.pet {
-    duration = 5000ms
-  }
-  retry.pet {
-    delay = 100ms
-    attempts = 2
-  }
-}
-
-
-logging.level {
-  "root": "WARN"
-  "ru.tinkoff.kora": "INFO"
-  "ru.tinkoff.kora.example": "INFO"
-  "ru.tinkoff.kora.application.graph.internal.loom.VirtualThreadExecutorHolder": "DEBUG"
-}
