+++ settings.gradle
@@ -31,6 +31,8 @@ include "kora-java-config-hocon"
 include "kora-java-config-yaml"
 include "kora-java-helloworld"
 include "kora-java-crud"
-include "kora-java-graalvm-crud"
+include "kora-java-graalvm-crud-jdbc"
+include "kora-java-graalvm-crud-r2dbc"
+include "kora-java-graalvm-crud-cassandra"
 include "kora-kotlin-helloworld"
 include "kora-kotlin-crud"
